requirements.txt
fire
ipdb
torchvision
visdom
tqdm
